.ipynb_checkpoints/requirements-checkpoint.txt
langchain-community
langchain
openai
tiktoken
sentence-transformers
qdrant-client
cohere